Move number param
